Finish abstract gateway implementation for backend gateways
