+++ packages/eth-edge/src/gateways/alchemy_node.gateway.ts
@@ -1,14 +1,12 @@
-import { request, Method, RequestFunc } from "@cryptify/common/src/helpers/request";
 import { Injectable } from "@nestjs/common";
 import { ConfigService } from "@nestjs/config";
+import {Gateway} from "@cryptify/common/src/gateway/gateway";
+import {AbstractGateway, Method} from "@cryptify/common/src/gateway/abstract_gateway";
 
 @Injectable()
-export class AlchemyNodeGateway {
-    request: RequestFunc;
-
+export class AlchemyNodeGateway extends Gateway {
     constructor(private configService: ConfigService) {
-        const uri = `https://${configService.get<string>("ALCHEMY_HOST")}`;
-        this.request = request(uri);
+        super(`https://${configService.get<string>("ALCHEMY_HOST")}`);
     }
 
     async updateWebhookAddresses(addressesToAdd: string[], addressesToRemove: string[]): Promise<void> {
@@ -22,6 +20,6 @@ export class AlchemyNodeGateway {
             webhook_id: this.configService.get<string>("ALCHEMY_ADDRESS_ACTIVITY_WEBHOOK_ID"),
         };
 
-        return this.request(Method.PATCH, headers, path, body);
+        await this.request<void>(Method.PATCH, headers, path, body);
     }
 }
